GHC.Generics.Generically

Description

This module exports Generically and Generically newtypes meant to be used with GHC.Generics and DerivingVia.

These types are re-exported from GHC.Generics on base-4.17 and later, and defined here for older base versions.

Documentation

newtype Generically a

A type whose instances are defined generically, using the Generic representation.

Constructors

Generically a

newtype Generically1 f a

A datatype whose instances are defined generically, using the Generic representation. Generically1 is a higher-kinded version of Generically that uses Generic1.

Constructors

Generically1 (f a)